changed the InheritedClassConverter structure a bit to make it easier to read and so we can put the `where TDiscriminator : struct` back
