+++ source/Nevermore/Mapping/StandardTypeResolver.cs
@@ -3,12 +3,6 @@ using System.Data;
 
 namespace Nevermore.Mapping
 {
-    public abstract class InstanceTypeResolver
-    {
-        public abstract Type GetTypeFromInstance(object instance);
-        public abstract Func<IDataReader, Type> TypeResolverFromReader(Func<string, int> columnOrdinal);
-    }
-
     public class StandardTypeResolver : InstanceTypeResolver
     {
         readonly DocumentMap mapper;
